refactoring to make room for push evaluation in roper
